move show more btn to the shared
